vision: Use CameraServer.launch()
This changes the vision code to use wpilib.CameraServer.launch().
As part of this change, this requires using some form of IPC other than
the shared memory primitives provided by multiprocessing, so we use
NetworkTables as set up by the cscore.__main__ wrapper.

There is also a need for the properties exposed by Vision to be
documented.  As such, this also uses the doc argument in
robotpy-wpilib-utilities 2017.0.4 to document these.

NB: ntproperty has less overhead than tunable, we may wish to change.