profilefollower: Use queues, avoid using internals
This changes the queues to use an actual efficient queue data structure,
rather than using lists and popping the 0th item.

This also removes the unnecessary list holding the queues and puts them
directly as members of ProfileFollower, and changes the users of
ProfileFollower to check the executing variable rather than the length of
the queue is 0, to avoid depending on implementation details.

Also fix a bug where curves autonomous wouldn't actually engage the
gears state machine.